refactor attention and activation to own files
